Add S-1 GRD EW
