+++ .env.example
@@ -4,5 +4,8 @@ CLIENT_SECRET=
 # Env var INSTANCE_ID should be set to the ID of some instance in Sentinel Hub
 # Configurator, which contains all the layers listed below it.
 INSTANCE_ID=
+S1GRDIW_LAYER_ID=
+S1GRDEW_LAYER_ID=
 S2L2A_LAYER_ID=
-S1GRD_LAYER_ID=
+S3SLSTR_LAYER_ID=
+S3OLCI_LAYER_ID=
